requirements.txt
PySimpleGUI
matplotlib
seaborn
shap
pandas
numpy
joblib
scikit-learn
xgboost
Pillow
kagglehub
